migrate oauth apis to dedicated tools routes
