+++ apps/sim/app/api/tools/discord/servers/route.ts
