+++ apps/sim/app/api/tools/jira/issues/route.ts
