+++ apps/sim/app/api/tools/jira/projects/route.ts
