+++ apps/sim/tools/confluence/retrieve.ts
@@ -43,7 +43,7 @@ export const confluenceRetrieveTool: ToolConfig<
   request: {
     url: (params: ConfluenceRetrieveParams) => {
       // Instead of calling Confluence API directly, use your API route
-      return '/api/auth/oauth/confluence/page'
+      return '/api/tools/confluence/page'
     },
     method: 'POST',
     headers: (params: ConfluenceRetrieveParams) => {
